+++ CHANGELOG.md
@@ -16,7 +16,9 @@ This version includes **multiple breaking changes**. If you are upgrading from v
   - Sorting order issues when emails lack `receivedAt` headers.
 - IMAP: Fix `BINARY` fetch responses (#2940).
 - ACME: Allow requesting apex domain certificates.
-- Bootstrap: Accept RFC 6761 reserved TLDs during bootstrap.
+- Hostname issues: 
+  - Accept RFC 6761 reserved TLDs during bootstrap.
+  - Allow hostnames without TLDs in remote server settings.
 - Reverse proxy issues.
 - OSS builds.
 - DNS Updater: